# lib/api-zod/src/generated/types/dashboardSyncRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSyncRunState } from './dashboardSyncRunState';

export interface DashboardSyncRun {
  /** @nullable */
  runId: string | null;
  state: DashboardSyncRunState;
  completedSteps: number;
  totalSteps: number;
  /** @nullable */
  error: string | null;
}