# lib/api-zod/src/generated/types/dashboardSyncStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSyncRun } from './dashboardSyncRun';
import type { DashboardSyncStatusAccessRole } from './dashboardSyncStatusAccessRole';

export type DashboardSyncStatus = DashboardSyncRun & ({
  canSync: boolean;
  accessRole: DashboardSyncStatusAccessRole;
  roleLocked: boolean;
  snapshotAvailable: boolean;
  /** @nullable */
  syncedAt: string | null;
});